CentOS 6.5 works (requires AWS account keypair)
For some reason the access key id and secret access key did not correctly create the AMI for CentOS. For some reason when attempting to connect it would time out. I opted instead of using an exisiting keypair (instead of having it aut-generated - which I think is what happens with the two keys).

So now there is a requirement that the AWS keypair be downloaded and set to an environment variable location.