FIX for CentOS 6.5 to disable SELinux Context
Disabling the service like I was doing did not work correctly. That is now fixed and this should be the corret way to disable the SELinux context. However, this may not even be necessary as we are not using non-standard ports with this image. But to allow for this image to be used by the older training it needs to be present.