Added TDD Cookbook Development Course & RENAMED Cookbook
The cookbook no longer helps create only the Essentials course. So the name of the cookbooks 'essentials' was not accurate. Changing it to workstations. Started to redefine the recipes in the name of the training workstation that needs to be created. I also started to try and add better Test Kitchen support. Hoever, this is not working correctly because of a bug with Busser not working properly with the ChefDK version that I get installed.