Changing Setup Tactics for TDD Cookbook Development Packer
Talking with Adam Leff it seems like the better strategy is to install Chef through the Chef Omnibus installer. Then use the recipe to install ChefDK. That is what these changes entail. Instead of doing it with a simple execute I chose to pull in the chef-ingredient cookbook. This cookbook has some requirements that were pulled in.

This check-in has the TDD Cookbook Development recipe working both in Test Kitchen and Packer. A mighty task.