Adding Windows Node and Workstation Packer Scripts (WARNING)
The big warning here is that the workstation script does not work as a chef recipe or a powershell script. However, I want to still capture it to be able to communicate it with others that might be able to assist with solving the problem.

The node appears to be working correctly.