Add the Compliance Windows Node (WARNING)
I am adding the Packer file here to create this AMI. However, this packer script does not currently work and fails when attempting to install components through the Nordstrom setup script that is part of this pprocess. This may be fixed in a future version of Packer or I may explore some changes to the install script to make it work. Until then this packer script really does the job of showing it it would look if it worked. This may be more confusing for someone or a mistake. I'm sorry if this burns you.

In other news, the compliance-windows powershell script sets the state of the system up when you do want to create a new AMI. Well everything except for the WinRM script that also need to be run!