Adds README to explain more about this project
This project does quite a few things. Create linux AMI using cookbooks.
Create Windows AMI using powershell scripts. The AWS directory contains
an IAM Role policy and I wanted to explain how that is used. I wanted
to quickly explain that the README in the cookbooks is to make sure
that people would be aware that it is all about the linux system.
Similar to the README in the scripts section. And the kitchen template
as well.

Signed-off-by: Franklin Webber <franklin@chef.io>