Fix for SSH cookbook (fork) known_hosts
Alright, so we have to keep the cookbooks here for packer to work
with them. There was an issue with Chef 13 which is now being used
to converge the nodes. This caused a problem with the ssh cookbook.
Instead of grabbing the latest ssh cookbook I opted to make this
single change.

When it comes time to replace the ssh cookbook we will want to
override it completely anywys so I don't think that this will
cause any major problems.

Famous last words.

Signed-off-by: Franklin Webber <franklin@chef.io>