Updates the README to share the workflow
So the basic flow for all of these images is probably
about the same. If the files are the same we could
automate the entire process of creating an image and
getting a ssh'd in session.

Of course, if this base ami were to update all the
images based on it should also be updated. Which probably
could be accomplished when/if there was a build system.

Signed-off-by: Franklin Webber <franklin@chef.io>